pyproject.toml
[project]
name = "mnist-fpga"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pyserial>=3.5",
    "scikit-learn>=1.7.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
